Test and Production profiles, and minor fixes.
The build has been modified so the default is for a local-test friendly build that plays-back the SOPA Wikipedia homepage. The 'production' profile rebuilds the same config with the right parameters for the live system.

Also updated the HTML error page to contain links to other holdings. Fixed up the favicon. Experimented with JavaScript overrides (currently disabled).